requirements.txt
discord.py==2.2.0
Babel==2.9.1
filetype==1.0.7
psycopg2
numpy==1.19.3
Pillow==8.3.1
requests==2.26.0
